requirements.txt
opencv-python
mediapipe
PySide6
PyYAML
pyautogui
watchdog
